Move ContentQuestion and exception classes to separate files
