+++ dmcontent/errors.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+
+class ContentNotFoundError(Exception):
+    pass
+
+
+class QuestionNotFoundError(Exception):
+    pass
